pyproject.toml
[tool.poetry]
name = "popleave"
version = "0.1.0"
description = "Populates Leave Forms from the Command Line"
authors = ["Alex Drysdale <alexander.drysdale@wales.nhs.uk>"]
readme = "README.md"
license = "GPL-3.0-or-later"
include = ["popleave/template.docx"]

[tool.poetry.dependencies]
python = "^3.12"
python-docx = "^1.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
